+++ cmd/core/entities/material.go
@@ -7,7 +7,7 @@ type Material struct {
 	Name            string
 	Description     string
 	Price           float64
-	Stock           int
-	RepositionPoint int
+	Stock           float64
+	RepositionPoint float64
 	MaterialType    MaterialType
 }
